+++ src/Model/PowerFlow/Generator4Governor/Generator4Governor.hpp
@@ -151,52 +151,52 @@ namespace GridKit
 
     const ScalarT dotDelta() const
     {
-      return yp_[offsetGen_ + 0];
+      return yp_[static_cast<size_t>(offsetGen_ + 0)];
     }
 
     const ScalarT dotOmega() const
     {
-      return yp_[offsetGen_ + 1];
+      return yp_[static_cast<size_t>(offsetGen_ + 1)];
     }
 
     const ScalarT dotEdp() const
     {
-      return yp_[offsetGen_ + 2];
+      return yp_[static_cast<size_t>(offsetGen_ + 2)];
     }
 
     const ScalarT dotEqp() const
     {
-      return yp_[offsetGen_ + 3];
+      return yp_[static_cast<size_t>(offsetGen_ + 3)];
     }
 
     const ScalarT delta() const
     {
-      return y_[offsetGen_ + 0];
+      return y_[static_cast<size_t>(offsetGen_ + 0)];
     }
 
     const ScalarT omega() const
     {
-      return y_[offsetGen_ + 1];
+      return y_[static_cast<size_t>(offsetGen_ + 1)];
     }
 
     const ScalarT Edp() const
     {
-      return y_[offsetGen_ + 2];
+      return y_[static_cast<size_t>(offsetGen_ + 2)];
     }
 
     const ScalarT Eqp() const
     {
-      return y_[offsetGen_ + 3];
+      return y_[static_cast<size_t>(offsetGen_ + 3)];
     }
 
     const ScalarT Id() const
     {
-      return y_[offsetGen_ + 4];
+      return y_[static_cast<size_t>(offsetGen_ + 4)];
     }
 
     const ScalarT Iq() const
     {
-      return y_[offsetGen_ + 5];
+      return y_[static_cast<size_t>(offsetGen_ + 5)];
     }
 
     const ScalarT K() const
